+++ packages/components/Stack/src/index.ts
@@ -1,4 +1,6 @@
-export * from './StackItem/StackItem.types';
-export * from './StackItem/StackItem';
-export * from './Stack.types';
-export * from './Stack';
+export { stackItemName } from './StackItem/StackItem.types';
+export type { IStackItemProps, IStackItemTokens, IStackItemType } from './StackItem/StackItem.types';
+export { StackItem } from './StackItem/StackItem';
+export { stackName } from './Stack.types';
+export type { Alignment, IStackProps, IStackRenderData, IStackSlotProps, IStackStatics, IStackTokens, IStackType } from './Stack.types';
+export { Stack } from './Stack';