+++ packages/framework/immutable-merge/src/index.ts
@@ -1,1 +1,2 @@
-export * from './Merge';
+export { immutableMerge, immutableMergeCore, processImmutable } from './Merge';
+export type { BuiltinRecursionHandlers, CustomRecursionHandler, MergeOptions, RecursionHandler, RecursionOption } from './Merge';