+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,99 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Wed, 23 Mar 2022 17:24:05 GMT and should not be manually modified.
+This log was last generated on Mon, 04 Apr 2022 23:40:49 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.59.2
+
+Mon, 04 Apr 2022 23:40:49 GMT
+
+### Patches
+
+- Factor duplicate code and remove dependencies on deprecated components (patboyd@microsoft.com)
+
+## 0.59.1
+
+Mon, 04 Apr 2022 18:41:35 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/apple-theme to v0.9.15
+
+## 0.59.0
+
+Thu, 31 Mar 2022 21:43:56 GMT
+
+### Minor changes
+
+- Upgrade appium (ruaraki@microsoft.com)
+
+## 0.58.16
+
+Thu, 31 Mar 2022 07:27:47 GMT
+
+### Patches
+
+- Ban `export *` in index files for better tree-shakeability (4123478+tido64@users.noreply.github.com)
+- Bump @fluentui-react-native/android-theme to v0.5.6
+- Bump @fluentui-react-native/apple-theme to v0.9.14
+- Bump @fluentui-react-native/badge to v0.0.21
+- Bump @fluentui-react-native/button to v0.22.11
+- Bump @fluentui-react-native/default-theme to v0.9.13
+- Bump @fluentui-react-native/experimental-activity-indicator to v0.5.18
+- Bump @fluentui-react-native/experimental-avatar to v0.14.10
+- Bump @fluentui-react-native/experimental-button to v0.15.11
+- Bump @fluentui-react-native/experimental-checkbox to v0.10.8
+- Bump @fluentui-react-native/experimental-drawer to v0.1.1
+- Bump @fluentui-react-native/experimental-expander to v0.3.20
+- Bump @fluentui-react-native/experimental-menu-button to v0.3.23
+- Bump @fluentui-react-native/experimental-shimmer to v0.6.22
+- Bump @fluentui-react-native/experimental-tabs to v0.4.32
+- Bump @fluentui-react-native/experimental-text to v0.7.19
+- Bump @fluentui-react-native/framework to v0.7.19
+- Bump @fluentui-react-native/icon to v0.11.14
+- Bump @fluentui-react-native/interactive-hooks to v0.14.6
+- Bump @fluentui-react-native/stack to v0.6.19
+- Bump @fluentui-react-native/theme to v0.6.6
+- Bump @fluentui-react-native/theme-types to v0.14.1
+- Bump @fluentui-react-native/themed-stylesheet to v1.3.3
+- Bump @fluentui-react-native/theming-utils to v0.9.2
+- Bump @fluentui-react-native/win32-theme to v0.14.15
+- Bump @fluentui/react-native to v0.28.1
+
+## 0.58.15
+
+Wed, 30 Mar 2022 16:31:51 GMT
+
+### Patches
+
+- Switch tests to use core package for Button (ruaraki@microsoft.com)
+- Bump @fluentui/react-native to v0.28.0
+
+## 0.58.14
+
+Mon, 28 Mar 2022 15:02:37 GMT
+
+### Patches
+
+- temp fix e2e testing error for menuButton (email not defined)
+- Bump @fluentui-react-native/badge to v0.0.20
+- Bump @fluentui-react-native/button to v0.22.10
+- Bump @fluentui-react-native/experimental-button to v0.15.10
+- Bump @fluentui-react-native/experimental-checkbox to v0.10.7
+- Bump @fluentui-react-native/experimental-drawer to v0.1.0
+- Bump @fluentui-react-native/experimental-menu-button to v0.3.22
+- Bump @fluentui-react-native/experimental-tabs to v0.4.31
+- Bump @fluentui-react-native/interactive-hooks to v0.14.5
+- Bump @fluentui/react-native to v0.27.44
+
+## 0.58.13
+
+Fri, 25 Mar 2022 20:06:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-checkbox to v0.10.6
 
 ## 0.58.12