+++ packages/components/MenuButton/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @fluentui-react-native/menu-button
 
-This log was last generated on Wed, 23 Mar 2022 17:24:05 GMT and should not be manually modified.
+This log was last generated on Thu, 31 Mar 2022 07:27:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.44
+
+Thu, 31 Mar 2022 07:27:47 GMT
+
+### Patches
+
+- Ban `export *` in index files for better tree-shakeability (4123478+tido64@users.noreply.github.com)
+- Bump @fluentui-react-native/button to v0.22.11
+- Bump @fluentui-react-native/component-cache to v1.3.2
+- Bump @fluentui-react-native/contextual-menu to v0.16.15
+- Bump @fluentui-react-native/icon to v0.11.14
+- Bump @fluentui-react-native/interactive-hooks to v0.14.6
+- Bump @fluentui-react-native/tokens to v0.11.9
+- Bump @uifabricshared/foundation-composable to v0.10.3
+- Bump @uifabricshared/foundation-compose to v1.11.13
+- Bump @uifabricshared/foundation-settings to v0.11.3
+
+## 0.7.43
+
+Mon, 28 Mar 2022 15:02:37 GMT
+
+### Patches
+
+- temp fix e2e testing error for menuButton (email not defined)
+- Bump @fluentui-react-native/button to v0.22.10
+- Bump @fluentui-react-native/contextual-menu to v0.16.14
+- Bump @fluentui-react-native/interactive-hooks to v0.14.5
 
 ## 0.7.42