+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,29 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Wed, 23 Mar 2022 17:24:05 GMT and should not be manually modified.
+This log was last generated on Thu, 31 Mar 2022 07:27:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.23
+
+Thu, 31 Mar 2022 07:27:47 GMT
+
+### Patches
+
+- Ban `export *` in index files for better tree-shakeability (4123478+tido64@users.noreply.github.com)
+- Bump @fluentui-react-native/contextual-menu to v0.16.15
+- Bump @fluentui-react-native/experimental-button to v0.15.11
+- Bump @fluentui-react-native/framework to v0.7.19
+- Bump @fluentui-react-native/tokens to v0.11.9
+
+## 0.3.22
+
+Mon, 28 Mar 2022 15:02:37 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.16.14
+- Bump @fluentui-react-native/experimental-button to v0.15.10
 
 ## 0.3.21